738. Missing deadkeys in UTF-8 compose table (Mike Fabian).  737. Build fix for SPARC (Thorsten Kukuk).  736. Fixed integer address value in vesa driver to be large enough       to hold a 64 bit memory address (Andreas Schwab).  735. Disabled RENDER accel on Matrox when used in multihead mode       (Stefan Dirsch).  734. Added FireGL 8700/8800 to the list of supported cards or RADEON       driver (Stefan Dirsch).  733. Added -mcmodel=kernel to DRM Makefile for x86_64 (Stefan Dirsch).  732. Added vesa and fbdev driver to x86_64, removed nsc driver for ia64 build       (Stefan Dirsch).

+++ programs/Xserver/hw/xfree86/drivers/vesa/vesa.h
@@ -98,7 +98,7 @@ typedef struct _VESARec
     miBankInfoRec bank;
     int curBank, bankSwitchWindowB;
     CARD16 maxBytesPerScanline;
-    int mapPhys, mapOff, mapSize;	/* video memory */
+    unsigned long mapPhys, mapOff, mapSize;	/* video memory */
     void *base, *VGAbase;
     CARD8 *state, *pstate;	/* SVGA state */
     int statePage, stateSize, stateMode;
